requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
boto3==1.35.0
pillow==10.4.0
pytesseract==0.3.13
werkzeug==3.0.3
